% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2int.R
\name{cat2int}
\alias{cat2int}
\title{Parse integer encoding for category variable in SQL}
\usage{
cat2int(feature = "feature", table = "table")
}
\description{
Parse integer encoding for category variable in SQL
}
