% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ks_sql.R
\name{parse_ks_sql}
\alias{parse_ks_sql}
\title{Parse SQL Code for KS test}
\usage{
parse_ks_sql(verbose = 0, y_hat = "pred", y = "target",
  table = "opd.test_pred_table")
}
\description{
Parse SQL Code for KS test
}
\examples{
\dontrun(parse_ks_sql(verbose = 0))
\dontrun(parse_ks_sql(verbose = 1))
\dontrun(parse_ks_sql(verbose = 2))
}
