% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_sql_result.R
\name{paste_sql_result}
\alias{paste_sql_result}
\title{Paste Markdown result from SQL query.}
\usage{
paste_sql_result(sql_text)
}
\description{
The output is formatted in markdown, os it is easy to share output at the GitHub or GitLab.
}
