% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_impala.R
\name{run_impala}
\alias{run_impala}
\title{Run impala from script}
\usage{
run_impala(file_path,
  log_path = here::here("output/caiweina/running-log.md"))
}
\arguments{
\item{file_path}{the file path of SQL content.}

\item{log_path}{the path to save log content.}
}
\description{
Run impala from script
}
