% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2int.R
\name{cat2int}
\alias{cat2int}
\title{Parse integer encoding for category variable in SQL}
\usage{
cat2int(feature = "feature", table = "table")
}
\arguments{
\item{feature}{Feature name.}

\item{table}{data.frame.}
}
\description{
Parse integer encoding for category variable in SQL
}
\examples{
\dontrun{library(RODBC)}
\dontrun{impala <- odbcConnect("Impala")}
\dontrun{output_cat2int <- cat2int(feature = "cyl", table = "opd.sqlsave_test_ljx")}
}
