% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_sql_result.R
\name{paste_sql_result}
\alias{paste_sql_result}
\title{Paste Markdown result from SQL query.}
\usage{
paste_sql_result(sql_text)
}
\arguments{
\item{sql_text}{The SQL text.}
}
\description{
The output is formatted in markdown, os it is easy to share output at the GitHub or GitLab.
}
\examples{
\dontrun{output_paste_sql_result1 <- paste_sql_result("describe opd.sqlsave_test_ljx")}
\dontrun{output_paste_sql_result2 <- read_rds("output/output_paste_sql_result2.rds")}
}
