% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_regex.R
\name{return_regex}
\alias{return_regex}
\title{Return Seemingly Regular Expressions.}
\usage{
return_regex(input = row.names(datasets::mtcars),
  verbose = c("recommend", "all", "lazy"))
}
\arguments{
\item{input}{The text list.}

\item{verbose}{the output style, by default, 'recommend', 'all' and 'lazy'.}
}
\description{
Return Seemingly Regular Expressions.
}
\examples{
inputs <- row.names(mtcars)
inputs \%>\% head
# remotes::install_github("daranzolin/inferregex")
return_regex(inputs, verbose = "recommend")
return_regex(inputs, verbose = "all")
return_regex(inputs, verbose = "lazy")
}
