% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_regex.R
\name{return_regex}
\alias{return_regex}
\title{Return Seemingly Regular Expressions.}
\usage{
return_regex(input = row.names(mtcars), verbose = c("recommend", "all",
  "lazy"))
}
\description{
Return Seemingly Regular Expressions.
}
