% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_ks.R
\name{rolling_ks}
\alias{rolling_ks}
\title{Parse SQL Code for rolling KS test}
\usage{
rolling_ks(y_hat = "pred", y = "target",
  table = "opd.test_pred_table", time_variable = "inserttime")
}
\description{
Parse SQL Code for rolling KS test
}
